moved manage button to the separate component
