+++ src/renderer/components/manageButton/ManageButton.styles.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+import { Button as UIButton } from '../uielements/button'
+
+export const ManageButton = styled(UIButton).attrs({
+  typevalue: 'outline',
+  round: 'true'
+})``
