+++ src/renderer/components/manageButton/index.ts
@@ -1,0 +1,1 @@
+export * from './ManageButton'
